tpmutil: seperate OpenTPM into emulion and non-emulation via build tag
to prohibit the inclusion of netowrk packages seperate the build of
tpmutil via buildtag 'emu' into a filesystem based and a socket based
version.
